Client 2958 (#2965)
* wip

* Add CalloutsPageLayout component

* Fix callouts data update on move card

* address pr comments
