+++ src/domain/challenge/hub/routing/HubRoute.tsx
@@ -11,10 +11,10 @@ import { Error404, PageProps } from '../../../../pages';
 import HubChallengesPage from '../pages/HubChallengesPage';
 import { routes } from '../routes/hubRoutes';
 import { EntityPageLayoutHolder } from '../../../shared/layout/PageLayout';
-import CalloutsPage from '../../../collaboration/callout/CalloutsPage';
 import CalloutRoute from '../../../collaboration/callout/routing/CalloutRoute';
 import HubContextPage from '../pages/HubContextPage';
 import HubDashboardPage from '../pages/HubDashboardPage';
+import CalloutsPage from '../../../collaboration/callout/CalloutsPage';
 
 export const HubRoute: FC<PageProps> = ({ paths: _paths }) => {
   const { displayName } = useHub();
@@ -46,7 +46,7 @@ export const HubRoute: FC<PageProps> = ({ paths: _paths }) => {
         />
         <Route
           path={`${routes.Explore}/callouts/:${nameOfUrl.calloutNameId}/*`}
-          element={<CalloutRoute parentPagePath={`${resolved.pathname}/${routes.Explore}`} entityTypeName={'hub'} />}
+          element={<CalloutRoute parentPagePath={`${resolved.pathname}/${routes.Explore}`} entityTypeName="hub" />}
         />
       </Route>
       <Route path={'apply'} element={<ApplyRoute paths={currentPaths} type={ApplicationTypeEnum.hub} />} />
