+++ src/domain/challenge/opportunity/routes/OpportunityRoute.tsx
@@ -8,10 +8,10 @@ import OpportunityAgreementsPage from '../pages/OpportunityAgreementsPage';
 import { nameOfUrl } from '../../../../core/routing/url-params';
 import { EntityPageLayoutHolder } from '../../../shared/layout/PageLayout';
 import { routes } from './opportunityRoutes';
-import CalloutsPage from '../../../collaboration/callout/CalloutsPage';
 import CalloutRoute from '../../../collaboration/callout/routing/CalloutRoute';
 import OpportunityContextPage from '../pages/OpportunityContextPage';
 import OpportunityDashboardPage from '../pages/OpportunityDashboardPage';
+import CalloutsPage from '../../../collaboration/callout/CalloutsPage';
 
 interface OpportunityRootProps extends PageProps {}
