+++ src/domain/collaboration/aspect/views/AspectRoute.tsx
@@ -1,6 +1,8 @@
 import React, { FC } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
+import { EntityTypeName } from '../../../shared/layout/PageLayout/SimplePageLayout';
 import useBackToParentPage from '../../../shared/utils/useBackToParentPage';
+import CalloutsPage from '../../callout/CalloutsPage';
 import AspectDashboardPage from '../pages/AspectDashboardPage';
 import AspectSettingsPage from '../pages/AspectSettingsPage';
 import AspectSharePage from '../pages/AspectSharePage';
@@ -9,9 +11,10 @@ import { AspectLayoutHolder } from './AspectLayoutWithOutlet';
 
 export interface AspectRouteProps {
   parentPagePath: string;
+  entityTypeName: EntityTypeName;
 }
 
-const AspectRoute: FC<AspectRouteProps> = ({ parentPagePath }) => {
+const AspectRoute: FC<AspectRouteProps> = ({ parentPagePath, entityTypeName }) => {
   const [backToExplore] = useBackToParentPage(parentPagePath, { keepScroll: true });
   const onClose = () => backToExplore();
 
@@ -19,9 +22,30 @@ const AspectRoute: FC<AspectRouteProps> = ({ parentPagePath }) => {
     <Routes>
       <Route path="/" element={<AspectLayoutHolder />}>
         <Route index element={<Navigate replace to={AspectDialogSection.Dashboard} state={{ keepScroll: true }} />} />
-        <Route path={AspectDialogSection.Dashboard} element={<AspectDashboardPage onClose={onClose} />} />
-        <Route path={AspectDialogSection.Share} element={<AspectSharePage onClose={onClose} />} />
-        <Route path={AspectDialogSection.Settings} element={<AspectSettingsPage onClose={onClose} />} />
+        <Route
+          path={AspectDialogSection.Dashboard}
+          element={
+            <CalloutsPage rootUrl={parentPagePath} entityTypeName={entityTypeName}>
+              <AspectDashboardPage onClose={onClose} />
+            </CalloutsPage>
+          }
+        />
+        <Route
+          path={AspectDialogSection.Share}
+          element={
+            <CalloutsPage rootUrl={parentPagePath} entityTypeName={entityTypeName}>
+              <AspectSharePage onClose={onClose} />
+            </CalloutsPage>
+          }
+        />
+        <Route
+          path={AspectDialogSection.Settings}
+          element={
+            <CalloutsPage rootUrl={parentPagePath} entityTypeName={entityTypeName}>
+              <AspectSettingsPage onClose={onClose} />
+            </CalloutsPage>
+          }
+        />
       </Route>
     </Routes>
   );
