+++ src/domain/collaboration/callout/useCallouts.ts
@@ -1,7 +1,10 @@
 import { isChallengeId, isHubId, isOpportunityId, OptionalCoreEntityIds } from '../../shared/types/CoreEntityIds';
 import {
+  useChallengeCalloutsLazyQuery,
   useChallengeCalloutsQuery,
+  useHubCalloutsLazyQuery,
   useHubCalloutsQuery,
+  useOpportunityCalloutsLazyQuery,
   useOpportunityCalloutsQuery,
 } from '../../../hooks/generated/graphql';
 import {
@@ -87,6 +90,20 @@ const useCallouts = (params: OptionalCoreEntityIds) => {
     skip: !isOpportunityId(params),
   });
 
+  const [getHubCallouts] = useHubCalloutsLazyQuery({
+    variables: isHubId(params) ? params : (params as never),
+    fetchPolicy: 'cache-and-network',
+  });
+  const [getChallengeCallouts] = useChallengeCalloutsLazyQuery({
+    variables: isChallengeId(params) ? params : (params as never),
+    fetchPolicy: 'cache-and-network',
+  });
+
+  const [getOpportunityCallouts] = useOpportunityCalloutsLazyQuery({
+    variables: isOpportunityId(params) ? params : (params as never),
+    fetchPolicy: 'cache-and-network',
+  });
+
   const collaboration = (
     hubCalloutsData?.hub ??
     challengeCalloutsData?.hub.challenge ??
@@ -123,10 +140,17 @@ const useCallouts = (params: OptionalCoreEntityIds) => {
     } as TypedCallout;
   });
 
+  const reloadCallouts = isOpportunityId(params)
+    ? getOpportunityCallouts
+    : isChallengeId(params)
+    ? getChallengeCallouts
+    : getHubCallouts;
+
   return {
     callouts,
     canCreateCallout,
     loading: hubCalloutsLoading || challengeCalloutsLoading || opportunityCalloutsLoading,
+    reloadCallouts,
   };
 };
